Restructured folders, minor styling
