+++ public/stylesheets/style.css
@@ -86,7 +86,7 @@ h5.card-title span.bf {
 
 /* RECIPE CONTAINER */
 div.card-block.recipe-card {
-  background-color: #ffffff;
+  background-color: #e7e7e7;
 }
 
 .recipe-list-wrapper {
