+++ test/App-test.js
@@ -6,9 +6,9 @@ import clientapi from '../app/clientapi'
 import * as router from 'react-router'
 import anims from '../app/utils/ui'
 import sinonStubPromise from 'sinon-stub-promise'
-import SearchContainer from '../app/containers/SearchContainer'
+import SearchContainer from '../app/components/Search'
 //TODO: change MainContainer to App
-import MainContainer from '../app/containers/App'
+import MainContainer from '../app/App'
 import Preloader from '../app/components/Preloader'
 import Index from '../app/components/Index'
 import Dashboard from '../app/components/Dashboard'
