+++ test/IngredientContainer-test.js
@@ -4,7 +4,7 @@ import {mount, shallow} from 'enzyme'
 import {assert} from 'chai'
 import anims from '../app/utils/ui'
 import sinonStubPromise from 'sinon-stub-promise'
-import SearchContainer from '../app/containers/SearchContainer'
+import SearchContainer from '../app/components/Search'
 import IngredientContainer from '../app/containers/IngredientContainer'
 import MainContainer from '../app/containers/MainContainer'
 import Preloader from '../app/components/Preloader'
