+++ test/action-test.js
@@ -1,5 +1,5 @@
 import {assert} from 'chai'
-import * as types from '../app/actions'
+import * as types from '../app/redux/actions'
 import sinon from 'sinon'
 
 describe('preFetchRecipes', () => {
